feature/ORD-445 DB error handler
Introduces changes to the error handling process for errors generated
at the database level. Currently, the application uses asyncpg
as the primary database driver. To handle errors, a middleware has
been added that catches specific exceptions and returns
appropriate status codes.

Support for psycopg is also maintained, as the tests still rely
on this driver.